Track unresolved Provider references
